[feat]: 블랙홀 리사이클러뷰로 옮기기
